+++ src/Utils.tsx
@@ -7,7 +7,7 @@ const REGEX_COLOR = /^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$/;
  */
 export const assertIsColor = (color: string) => {
    if(!color.match(REGEX_COLOR))  {
-       throw Error("Color me be an hexadecimal color")
+       throw Error("Color me be an hexadecimal color with # (e.g #FF0000)")
    }
 }